x64 fixes (see details) fixed many warnings through proper static casts slight renames in reliable_channel fixed rectpack2D randomization_declaration.h - does using rng_seed_type and chosen unsigned for that constant_size_vector stores unsigned instead of std::size_t for portability byte readwrite uses unsigned by default for container size, for portability proper reinterpret casts for describe_fields.h fixed casting in shell.cpp - though not sure if correct fixed raw mouse input handling

+++ src/3rdparty/Box2D/Collision/b2DynamicTree.cpp
@@ -804,7 +804,7 @@ b2DynamicTree& b2DynamicTree::operator=(const b2DynamicTree& b)
 #endif
 #endif
 
-	m_nodes = (b2TreeNode*)b2Alloc(bytes);
+	m_nodes = (b2TreeNode*)b2Alloc(static_cast<int32>(bytes));
 	memcpy(m_nodes, b.m_nodes, bytes);
 
 	m_root = b.m_root;
