+++ src/augs/misc/declare_containers.h
@@ -4,6 +4,6 @@ namespace augs {
 	template <class Enum, class T, class = void>
 	class enum_associative_array;
 
-	template <class T, size_t, class = void>
+	template <class T, unsigned, class = void>
 	class constant_size_vector;
 }
